feat: best vote 예외처리
