+++ package.json
@@ -54,5 +54,6 @@
     "typescript": "^5",
     "typescript-eslint": "^8.32.1",
     "webpack": "^5.99.9"
-  }
+  },
+  "packageManager": "pnpm@10.14.0+sha512.ad27a79641b49c3e481a16a805baa71817a04bbe06a38d17e60e2eaee83f6a146c6a688125f5792e48dd5ba30e7da52a5cda4c3992b9ccf333f9ce223af84748"
 }
