+++ src/api/votes.ts
@@ -1,12 +1,16 @@
 import { CreateVoteData, MineVotesResponse } from '@/types/api/votes'
 import { authApi } from './instance/authApi'
 import { VoteCategory } from '@/types/_shared/vote'
+import { isAxiosError } from 'axios'
 
 export const fetchBestVote = async () => {
   try {
     const response = await authApi.get<BestVoteResponse>('/votes/best')
     return response.data
   } catch (error) {
+    if (isAxiosError(error) && error.response?.status === 404) {
+      return null
+    }
     throw error
   }
 }
